+++ react/admin/ReturnForm.tsx
@@ -977,7 +977,8 @@ class ReturnForm extends Component<any, any> {
             refundedShippingValue={refundedShippingValue}
             product={product}
             totalRefundAmount={request.refundedAmount}
-            productsValue={totalAmount}
+            // This is a quick (and ugly) fix for the issue where the refunded amount is not being displayed for admins.
+            productsValue={totalAmount || request.totalPrice}
           />
           <p className="mt7">
             <strong className="mr6">