+++ react/admin/ReturnsSettings.tsx
@@ -40,6 +40,13 @@ const messages = defineMessages({
   searchCategories: { id: 'settings.searchCategories' },
   excludedCategories: { id: 'settings.excludedCategories' },
   paymentMethodsLabel: { id: 'settings.paymentMethods_label' },
+  hidePaymentMethodSelection: { id: 'settings.hidePaymentMethodSelection' },
+  hidePaymentMethodSelection_checked: {
+    id: 'settings.hidePaymentMethodSelection_checked',
+  },
+  hidePaymentMethodSelection_unchecked: {
+    id: 'settings.hidePaymentMethodSelection_unchecked',
+  },
   returnOptionsLabel: { id: 'settings.returnOptions_label' },
   noCustomOptions: { id: 'settings.noCustomOptions' },
   noCustomOptionsHowTo: { id: 'settings.noCustomOptions_HowTo' },
@@ -179,6 +186,8 @@ class ReturnsSettings extends Component<any, any> {
           options: json[0].options || [],
           excludedCategories: JSON.parse(json[0].excludedCategories),
           enableOtherOption: json[0].enableOtherOption,
+          hidePaymentMethodSelection:
+            json[0].hidePaymentMethodSelection ?? false,
           loading: false,
           enablePickupPoints: json[0].enablePickupPoints,
         })
@@ -319,6 +328,7 @@ class ReturnsSettings extends Component<any, any> {
       options,
       enableOtherOption,
       enablePickupPoints,
+      hidePaymentMethodSelection,
     } = this.state
 
     console.log(enablePickupPoints, 'stateee')
@@ -346,6 +356,7 @@ class ReturnsSettings extends Component<any, any> {
     }
 
     if (
+      !hidePaymentMethodSelection &&
       !payments.paymentBank.checked &&
       !payments.paymentCard.checked &&
       !payments.paymentVoucher.checked
@@ -376,6 +387,7 @@ class ReturnsSettings extends Component<any, any> {
       options,
       enableOtherOption,
       enablePickupPoints,
+      hidePaymentMethodSelection,
       type: schemaTypes.settings,
     }
 
@@ -475,6 +487,7 @@ class ReturnsSettings extends Component<any, any> {
       isModalOpen,
       enableOtherOption,
       enablePickupPoints,
+      hidePaymentMethodSelection,
     } = this.state
 
     const { formatMessage } = this.props.intl
@@ -608,21 +621,49 @@ class ReturnsSettings extends Component<any, any> {
                   ) : null}
                 </div>
                 <Divider orientation="horizontal" />
-                <div className="flex flex-column w-100">
-                  <p className="f4 mb6">
-                    {formatMessage({ id: messages.paymentMethodsLabel.id })}
+                <div className="flex flex-column w-100 mb6">
+                  <p className="f5 mv4">
+                    {formatMessage({
+                      id: messages.hidePaymentMethodSelection.id,
+                    })}
                   </p>
-                  <CheckboxGroup
-                    name="simpleCheckboxGroup"
-                    label={formatMessage({ id: messages.all.id })}
-                    id="simple"
-                    value="simple"
-                    checkedMap={payments}
-                    onGroupChange={(newCheckedMap) => {
-                      this.setState({ payments: newCheckedMap })
-                    }}
+                  <Toggle
+                    label={
+                      hidePaymentMethodSelection
+                        ? formatMessage({
+                            id: messages.hidePaymentMethodSelection_checked.id,
+                          })
+                        : formatMessage({
+                            id: messages.hidePaymentMethodSelection_unchecked
+                              .id,
+                          })
+                    }
+                    semantic
+                    checked={hidePaymentMethodSelection}
+                    onChange={() =>
+                      this.setState((prevState) => ({
+                        hidePaymentMethodSelection:
+                          !prevState.hidePaymentMethodSelection,
+                      }))
+                    }
                   />
-
+                  {!hidePaymentMethodSelection && (
+                    <>
+                      <p className="f5 mb6">
+                        {formatMessage({ id: messages.paymentMethodsLabel.id })}
+                      </p>
+                      <CheckboxGroup
+                        name="simpleCheckboxGroup"
+                        label={formatMessage({ id: messages.all.id })}
+                        id="simple"
+                        value="simple"
+                        checkedMap={payments}
+                        onGroupChange={(newCheckedMap) => {
+                          this.setState({ payments: newCheckedMap })
+                        }}
+                      />
+                    </>
+                  )}
                   {errors.payments && (
                     <p className={`${styles.errorMessage}`}>
                       {errors.payments}