+++ react/components/RequestForm.tsx
@@ -10,6 +10,7 @@ import {
   Dropdown,
   Textarea,
   Toggle,
+  Spinner,
 } from 'vtex.styleguide'
 
 import NEAR_PICKUP_POINTS from '../graphql/nearestPickupPoints.gql'
@@ -93,6 +94,7 @@ const messages = defineMessages({
   formState: { id: 'returns.formState' },
   formZip: { id: 'returns.formZip' },
   formPaymentMethod: { id: 'returns.formPaymentMethod' },
+  defaultPaymentMethod: { id: 'returns.defaultPaymentMethod' },
   formIBAN: { id: 'returns.formIBAN' },
   formAccountHolder: { id: 'returns.formAccountHolder' },
   formNextStep: { id: 'returns.formNextStep' },
@@ -437,6 +439,89 @@ const RequestForm = (props) => {
               className={`w-100 f7 f6-xl fw4 c-muted-1 ttu ${styles.orderInfoHeaderOrderDateLabel}`}
             >
               {intl.formatMessage({ id: messages.orderDate.id })}
+            </div>
+            <div>
+              <table className={styles.tblProducts}>
+                <thead className={styles.tableThead}>
+                  <tr className={styles.tableTr}>
+                    <th className={styles.tableTh} />
+                    <th className={styles.tableTh}>
+                      {intl.formatMessage({ id: messages.thProduct.id })}
+                    </th>
+                    <th className={styles.tableTh}>
+                      {intl.formatMessage({ id: messages.thQuantity.id })}
+                    </th>
+                    <th className={styles.tableTh}>
+                      {intl.formatMessage({ id: messages.thReason.id })}
+                    </th>
+                    <th className={styles.tableTh}>
+                      {intl.formatMessage({ id: messages.condition.id })}
+                    </th>
+                  </tr>
+                </thead>
+                {!props.orderProducts.length ? (
+                  <Spinner />
+                ) : (
+                  <tbody className={styles.tableTbody}>
+                    {props.orderProducts.map((product: any) => (
+                      <tr
+                        key={`product${product.uniqueId}`}
+                        className={styles.tableTr}
+                      >
+                        <td
+                          className={`${styles.tableTd} ${styles.tableTdImage}`}
+                        >
+                          <img
+                            className={styles.imageCol}
+                            src={product.imageUrl}
+                            alt={product.name}
+                          />
+                        </td>
+                        <td className={`${styles.tableTd} ${styles.w350}`}>
+                          <a
+                            className={styles.productUrl}
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            href={product.detailUrl}
+                          >
+                            {product.name}
+                          </a>
+                        </td>
+                        <td
+                          className={`${styles.tableTd} ${styles.tableTdQuantity}`}
+                        >
+                          <Input
+                            suffix={`/${product.quantity}`}
+                            size="regular"
+                            type="number"
+                            value={product.selectedQuantity}
+                            onChange={(e) => {
+                              props.handleQuantity(product, e.target.value)
+                            }}
+                            max={product.quantity}
+                            min={0}
+                          />
+                        </td>
+                        <td
+                          className={`${styles.tableTd} ${styles.tableTdReason}`}
+                        >
+                          {() => renderReasonsDropdown(product)}
+                        </td>
+                        <td
+                          className={`${styles.tableTd} ${styles.tableTdReason}`}
+                        >
+                          {() => renderConditionDropdown(product)}
+                        </td>
+                      </tr>
+                    ))}
+                  </tbody>
+                )}
+              </table>
+              {props.errors.productQuantities ? (
+                <p className={styles.errorMessage}>
+                  {intl.formatMessage({ id: props.errors.productQuantities })}
+                </p>
+              ) : null}
             </div>
             <div
               className={`db pv0 f6 fw5 c-on-base f5-l ${styles.orderInfoHeaderOrderDate}`}